PWT-5 restoring DefaultConfig.php
